added more core concepts to system prompt and moved prompts to separate file
