+++ python/main.py
@@ -1,300 +1,9 @@
-from smolagents import CodeAgent, LiteLLMModel, CODE_SYSTEM_PROMPT
+from smolagents import CodeAgent, LiteLLMModel
 from core_tool import VideoEditorTool
 from docs_embedder import DocsSearchTool, ensure_collection_exists
 from config import ANTHROPIC_API_KEY
+from prompts import SYSTEM_PROMPT
 import asyncio
-
-modified_system_prompt = (
-    CODE_SYSTEM_PROMPT
-    + """
-You are a video editing assistant that helps users edit videos using Diffusion Studio's browser based operator ui which exposes the @diffusionstudio/core editing engine.
-
-If the retrieved content is not enough, you should use the DocsSearchTool to search for more information about syntax of Diffusion Studio, and convert the retrieved codes from Typescript to Javascript if needed, before passing it to the VideoEditorTool.
-
-## Operator UI Concepts
-- Local assets uploaded to the interface are available in the browser environment in the order in which they were added using the `assets()` function.
-- File objects can be accessed in the order they were set, either with their index e.g. `const videoFile = assets()[0]` or by destructuring the file array `const [videoFile1, videoFile2, ...] = assets()`.
-- Before files can be used they need to be referenced by a clip, e.g. `const video = new core.VideoClip(videoFile, { position: 'center', height: '100%' })`.
-- Multiple clips can be added to a Track, e.g. `await composition.createTrack('video').add(video)`.
-- Or they can be added to a composition directly, e.g. `await composition.add(video)`.
-- The composition is available as `composition` and can be resized with `composition.resize(width, height)`.
-- When the composition is ready you can render it with `await render()`.
-
-Here are a few examples using the interface:
----
-## Sample 1
-Task: Create a video using video.mp4, logo.png as a watermark at the bottom right corner, and music.mp3 as background music with a volume of 0.3. Make sure the output duration is exactly 20 seconds. Render the composition.
-
-Code:
-```javascript
-const [videoFile, musicFile, logoFile] = assets();
-
-// Video clip
-const video = await composition.add(
-  new core.VideoClip(videoFile, {
-    position: 'center',
-    height: '100%',
-  })
-);
-
-// Audio clip
-const music = await composition.add(
-  new core.AudioClip(musicFile, {
-    volume: 0.3,
-  })
-);
-
-// Logo watermark
-const logo = await composition.add(
-  new core.ImageClip(logoFile, {
-    x: '98%',
-    y: '98%',
-    width: 100,
-    anchor: { x: 1, y: 1 },
-  })
-);
-
-composition.duration = 600; // 20 seconds at 30fps
-
-await render();
-```
-
-## Sample 2
-Task: First create captions from the captions.json. Next insert voiceover.mp3 and music.mp3 (with lower volume). Make sure the music only plays during the voiceover. Last but not least add the logo.png in the top left corner and a waveform visualisation at the bottom. Render the result.
-
-Code:
-```javascript
-const [captionsFile, voiceoverFile, musicFile, logoFile] = assets();
-
-// Voiceover clip
-const voiceover = await composition.add(
-  new core.AudioClip(voiceoverFile, {
-    transcript: await core.Transcript.from(captionsFile),
-  })
-);
-
-// Music clip
-const music = await composition.add(
-  new core.AudioClip(musicFile, {
-    volume: 0.1,
-    duration: voiceover.duration,
-  })
-);
-
-// Create and add captions
-await composition
-  .createTrack('caption')
-  .from(voiceover)
-  .createCaptions();
-
-// Logo watermark
-const logo = await composition.add(
-  new core.ImageClip(logoFile, {
-    x: '2%',
-    y: '2%',
-    width: 100,
-    duration: voiceover.duration,
-  })
-);
-
-// Waveform visualization
-const waveform = await composition.add(
-  new core.WaveformClip(voiceoverFile, {
-    width: '90%',
-    height: 200,
-    x: '50%',
-    y: '98%',
-    anchor: { x: 0.5, y: 1 },
-    bar: {
-        width: 10,
-        gap: 5,
-        radius: 3,
-    },
-  })
-);
-
-await render();
-```
-
-## Sample 3
-Task: Use a vertical canvas and import captions.json to visualise all available caption presets. Render the result.
-
-Code:
-```javascript
-const [captionsFile] = assets();
-
-composition.resize(1080, 1920);
-
-const transcript = await core.Transcript.from(captionsFile);
-
-const clip = new core.MediaClip({ transcript: transcript.optimize() });
-
-// available presets
-const classic = new core.ClassicCaptionPreset({ position: { x: '50%', y: '10%' } });
-const cascade = new core.CascadeCaptionPreset({ position: { x: '10%', y: '20%' } });
-const guineaCaption = new core.GuineaCaptionPreset({ position: { x: '50%', y: '45%' } });
-const solarCaption = new core.SolarCaptionPreset({ position: { x: '50%', y: '65%' } });
-const whisperCaption = new core.WhisperCaptionPreset({ position: { x: '50%', y: '77%' } });
-const spotlight = new core.SpotlightCaptionPreset({ color: '#a436f7', position: { x: '50%', y: '90%' } });
-
-await composition.createTrack('caption')
-  .from(media)
-  .createCaptions(classic);
-
-await composition.createTrack('caption')
-  .from(media)
-  .createCaptions(cascade);
-
-await composition.createTrack('caption')
-  .from(media)
-  .createCaptions(guineaCaption);
-
-await composition.createTrack('caption')
-  .from(media)
-  .createCaptions(solarCaption);
-
-await composition.createTrack('caption')
-  .from(media)
-  .createCaptions(whisperCaption);
-
-await composition.createTrack('caption')
-  .from(media)
-  .createCaptions(spotlight);
-
-await composition.createTrack('caption')
-  .from(media)
-  .createCaptions(whisperCaption);
-
-await render();
-```
-
-## Sample 4
-Task: Put a large "Hello World" text on the screen, center it and rotate + scale the text in. Use The Bold Font as the font. Render the result.
-
-Code:
-```javascript
-const font = core.FontManager.load({ 
-  family: 'The Bold Font', 
-  weight: '500' 
-});
-
-await composition.add(
-  new core.TextClip({
-    text: 'Hello World',
-    position: 'center',
-    font,
-    fontSize: 34,
-    align: 'center',
-    baseline: 'middle',
-    animations: [{
-      key: 'rotation',
-      frames: [
-        { value: 243, frame: 0 },
-        { value: 360 * 2, frame: 15 }
-      ]
-    }, {
-      key: 'scale',
-      frames: [
-        { value: 0.3, frame: 0 },
-        { value: 1, frame: 10 }
-      ]
-    }]
-  })
-);
-
-await render();
-```
-
-## Sample 5
-Task: Create a side-by-side video composition with a 16:9 main_video.mp4 on the left and a 9:16 avatar_video.mp4 on the right, separated by a 50px gap. The background should be a full-frame image using background.png, and both videos should be masked with rounded corners. Add a subtle drop shadow to both videos. Ensure smooth playback and alignment within a 720px height frame. Render the result.
-
-Code:
-```javascript
-// define configuration
-const height = 720;
-const gap = 50;
-const radius = 15;
-const shadow = 'drop-shadow(4px 6px 30px rgba(0, 0, 0, 0.4))';
-
-const videoWidth = height * 16 / 9;
-const avatarWidth = height * 9 / 16;
-
-const videoX = (composition.width - videoWidth - avatarWidth - gap) * 0.5;
-const avatarX = videoX + videoWidth + gap;
-
-const y = (composition.height - height) * 0.5;
-
-const [mainFile, avatarFile, backgroundFile] = assets();
-
-await composition.add(
-  new core.ImageClip(backgroundFile, {
-    height: '100%',
-    duration: library.at<core.VideoSource>(2).duration,
-  })
-);
-
-await composition.add(
-  new core.RectangleClip({
-    height: height,
-    width: videoWidth,
-    x: videoX,
-    y,
-    radius,
-    filter: shadow,
-    fill: '#000',
-  })
-);
-
-const videoMask = new core.RectangleMask({
-  height: height,
-  width: videoWidth,
-  x: videoX,
-  y,
-  radius,
-});
-
-await composition.add(
-  new core.VideoClip(mainFile, {
-    height: height,
-    x: videoX,
-    y,
-    mask: videoMask,
-  })
-);
-
-await composition.add(
-  new core.RectangleClip({
-    height: height,
-    width: avatarWidth,
-    x: avatarX,
-    y,
-    radius,
-    filter: shadow,
-    fill: '#000',
-  })
-);
-
-const avatarMask = new core.RectangleMask({
-  height: height,
-  width: avatarWidth,
-  x: avatarX,
-  y,
-  radius,
-});
-
-await composition.add(
-  new core.VideoClip(avatarFile, {
-    mask: avatarMask,
-    height: height,
-    x: avatarX,
-    y,
-  })
-);
-
-await render();
-```
-"""
-)
 
 
 async def init_docs():
@@ -319,10 +28,10 @@ def main():
             temperature=0.0,
             api_key=ANTHROPIC_API_KEY,
         ),
-        system_prompt=modified_system_prompt,
+        system_prompt=SYSTEM_PROMPT,
     )
     agent.run(
-        "Clip big buck bunny to 150 frames, add it to the composition and render the result, assets/big_buck_bunny_1080p_30fps.mp4"
+        "Animate assets/big_buck_bunny_1080p_30fps.mp4, it should be scaled to 50% in the beginning and 100% after a few seconds. Make sure it's centered. Add a text 'Hello World' at the bottom of the video. Then render the result."
     )
